LayoutTests/js/Promise-then-in-workers-expected.txt

[Worker] Test Promise.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: resources/Promise-then-in-workers.js
PASS [Worker] thisInInit === firstPromise is true
PASS [Worker] firstPromise instanceof Promise is true
PASS [Worker] secondPromise instanceof Promise is true
PASS [Worker] firstPromise.then(null) threw exception TypeError: Expected function or undefined as as first argument.
PASS [Worker] firstPromise.then(undefined, null) threw exception TypeError: Expected function or undefined as as second argument.
PASS [Worker] firstPromise.then(37) threw exception TypeError: Expected function or undefined as as first argument.
PASS [Worker] thisInFulfillCallback === firstPromise is false
PASS [Worker] thisInFulfillCallback === secondPromise is true
PASS [Worker] result is "hello"
PASS successfullyParsed is true

TEST COMPLETE
